Fixes #7258: Response was sending HTML content type when formatter was set to JSON or XML, nulls were handled wrong
